Move DefBuffer and DefPackage to the type2 module
Not sure why they were in term_object before, as they're both Type 2 opcodes.
